+++ src/components/image.ts
@@ -3,17 +3,17 @@ import { ComponentId } from '../common'
 export class Image implements Component {
   id: ComponentId
   src: string
-  alt: string
-  width: number
-  height: number
-  url: string
+  alt?: string
+  width?: number
+  height?: number
+  url?: string
 
   constructor(
     src: string,
-    alt: string,
-    width: number,
-    height: number,
-    url: string,
+    alt?: string,
+    width?: number,
+    height?: number,
+    url?: string,
   ) {
     this.id = ComponentId.Image
     this.src = src