+++ src/components/viewInBrowser.ts
@@ -2,9 +2,9 @@ import { ComponentId } from '../common'
 
 export class ViewInBrowser implements Component {
   id: ComponentId
-  text: string
+  text?: string
 
-  constructor(text: string) {
+  constructor(text?: string) {
     this.id = ComponentId.ViewInBrowser
     this.text = text
   }