+++ src/content.ts
@@ -75,9 +75,7 @@ class ContentBuilder {
     height?: number,
     alt?: string,
   ): this {
-    return this.push(
-      new Image(src, alt ?? '', width ?? 0, height ?? 0, url ?? ''),
-    )
+    return this.push(new Image(src, alt, width, height, url))
   }
 
   link(text: string, url: string): this {
@@ -96,7 +94,7 @@ class ContentBuilder {
     return this.push(new Text(text))
   }
 
-  viewInBrowser(text: string): this {
+  viewInBrowser(text?: string): this {
     return this.push(new ViewInBrowser(text))
   }